Greatly improve the synchronisation process!
Now, fetched files are not fetched more than once a day,
as they are cached in a directory for the day in question.
Soon, this will be possible to override. If a given file
is not present, it will not be fetched again.

This is a good improvement as the timestamp is now pretty
understandable if one is looking in the directory.